+++ bin/TA_dmarc_rh_dmarc_imap.py
@@ -64,6 +64,30 @@ fields = [
         default=True,
         validator=None
     ), 
+    field.RestField(
+        'validate_dkim',
+        required=False,
+        encrypted=False,
+        default=None,
+        validator=None
+    ), 
+    field.RestField(
+        'imap_mailbox',
+        required=True,
+        encrypted=False,
+        default='INBOX',
+        validator=validator.String(
+            min_len=0, 
+            max_len=8192, 
+        )
+    ), 
+    field.RestField(
+        'output_format',
+        required=True,
+        encrypted=False,
+        default='json',
+        validator=None
+    ), 
 
     field.RestField(
         'disabled',