+++ bin/input_module_dmarc_imap.py
@@ -15,12 +15,14 @@ def validate_input(helper, definition):
     """Implement your own validation logic to validate the input stanza configurations"""
 
     opt_imap_server    = definition.parameters.get("imap_server", None)
+    opt_imap_mailbox   = definition.parameters.get("imap_mailbox", None)
     opt_use_ssl        = True
     opt_global_account = definition.parameters.get('global_account', None)
+    opt_validate_dkim  = definition.parameters.get('validate_dkim', None)
 
     try:
         tmp_dir = create_tmp_dir(helper)
-        i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account)
+        i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account, opt_imap_mailbox, opt_validate_dkim)
         i2d.get_imap_connectivity()
     finally:
         remove_tmp_dir(helper, tmp_dir)
@@ -29,17 +31,19 @@ def collect_events(helper, ew):
     """Implement your data collection logic here """
 
     opt_imap_server    = helper.get_arg("imap_server")
+    opt_imap_mailbox   = helper.get_arg("imap_mailbox")
     opt_use_ssl        = True
     opt_global_account = helper.get_arg('global_account')
     opt_resolve_ip     = helper.get_arg('resolve_ip')
     opt_validate_xml   = helper.get_arg('validate_xml')
-    opt_output_format  = "kv"
+    opt_validate_dkim  = helper.get_arg('validate_dkim')
+    opt_output_format  = helper.get_arg('output_format')
 
     loglevel   = helper.get_log_level()
     helper.set_log_level(loglevel)
 
     tmp_dir = create_tmp_dir(helper)
-    i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account)
+    i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account, opt_imap_mailbox, opt_validate_dkim)
     try:
         filelist = i2d.process_incoming()
         if len(filelist)>0:
@@ -48,3 +52,17 @@ def collect_events(helper, ew):
                 d2s.process_incoming()
     finally:
         remove_tmp_dir(helper, tmp_dir)
+
+# PSEUDOCODE for refactor:
+#
+# mailbox = DMARCMailbox(imap, ssl, account)
+# for uid, message in mailbox.get_dmarc_messages()
+#     mail = DMARCMail(message)
+#     dkimvrfy = mail.dkim_verify()
+#     for file in mail.get_dmarc_attachments()
+#         rua = DMARCfile(file)
+#         res_xmlvalidation = rua.get_xml_validation()
+#         res_feedback = rua.get_rua_feedback()
+#         event = DMARCEvent(res_feedback, res_xmlvalidation, dkimvrfy)
+#         event.save_event()
+#     mailbox.save_checkpoint(uid)