+++ bin/input_module_dmarc_pop3.py
@@ -1,6 +1,6 @@
 # encoding = utf-8
 
-from dmarc.imap2dir import Imap2Dir
+from dmarc.pop2dir import Pop2Dir
 from dmarc.dir2splunk import Dir2Splunk
 from dmarc.helper import create_tmp_dir
 from dmarc.helper import remove_tmp_dir
@@ -14,36 +14,36 @@ from dmarc.helper import remove_tmp_dir
 def validate_input(helper, definition):
     """Implement your own validation logic to validate the input stanza configurations"""
 
-    opt_imap_server    = definition.parameters.get("imap_server", None)
-    opt_imap_mailbox   = definition.parameters.get("imap_mailbox", None)
+    opt_pop3_server    = definition.parameters.get("pop3_server", None)
     opt_use_ssl        = True
     opt_global_account = definition.parameters.get('global_account', None)
+    opt_validate_dkim  = definition.parameters.get('validate_dkim', None)
 
     try:
         tmp_dir = create_tmp_dir(helper)
-        i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account, opt_imap_mailbox)
-        i2d.get_imap_connectivity()
+        p2d = Pop2Dir(helper, opt_pop3_server, tmp_dir, opt_use_ssl, opt_global_account, opt_validate_dkim)
+        p2d.get_pop3_connectivity()
     finally:
         remove_tmp_dir(helper, tmp_dir)
 
 def collect_events(helper, ew):
     """Implement your data collection logic here """
 
-    opt_imap_server    = helper.get_arg("imap_server")
-    opt_imap_mailbox   = helper.get_arg("imap_mailbox")
+    opt_pop3_server    = helper.get_arg("pop3_server")
     opt_use_ssl        = True
     opt_global_account = helper.get_arg('global_account')
     opt_resolve_ip     = helper.get_arg('resolve_ip')
     opt_validate_xml   = helper.get_arg('validate_xml')
-    opt_output_format  = "json"
+    opt_validate_dkim  = helper.get_arg('validate_dkim')
+    opt_output_format  = helper.get_arg('output_format')
 
     loglevel   = helper.get_log_level()
     helper.set_log_level(loglevel)
 
     tmp_dir = create_tmp_dir(helper)
-    i2d = Imap2Dir(helper, opt_imap_server, tmp_dir, opt_use_ssl, opt_global_account, opt_imap_mailbox)
+    p2d = Pop2Dir(helper, opt_pop3_server, tmp_dir, opt_use_ssl, opt_global_account, opt_validate_dkim)
     try:
-        filelist = i2d.process_incoming()
+        filelist = p2d.process_incoming()
         if len(filelist)>0:
             d2s = Dir2Splunk(ew, helper, tmp_dir, 0, opt_resolve_ip, opt_validate_xml, opt_output_format, False)
             if d2s.check_dir():
@@ -56,12 +56,11 @@ def collect_events(helper, ew):
 # mailbox = DMARCMailbox(imap, ssl, account)
 # for uid, message in mailbox.get_dmarc_messages()
 #     mail = DMARCMail(message)
-#     dkimvrfy = DMARCMail.dkim_verify()
+#     dkimvrfy = mail.dkim_verify()
 #     for file in mail.get_dmarc_attachments()
 #         rua = DMARCfile(file)
 #         res_xmlvalidation = rua.get_xml_validation()
 #         res_feedback = rua.get_rua_feedback()
-#         res_dkimvrfy = dkimvrfy
-#         event = DMARCEvent(res_feedback, res_xmlvalidation, res_dkimvrfy)
+#         event = DMARCEvent(res_feedback, res_xmlvalidation, dkimvrfy)
 #         event.save_event()
 #     mailbox.save_checkpoint(uid)